wander-66 rewrite user profile page and slowly start to change styling to neobrutal
