+++ package.json
@@ -14,6 +14,8 @@
     "@heroicons/react": "^2.2.0",
     "@hookform/resolvers": "^3.9.1",
     "@prisma/client": "^5.21.1",
+    "@radix-ui/react-label": "^2.1.1",
+    "@radix-ui/react-select": "^2.1.5",
     "@radix-ui/react-slot": "^1.1.1",
     "@react-google-maps/api": "^2.20.3",
     "@stripe/react-stripe-js": "^3.0.0",
